requirements.txt
pandas
openpyxl
python-dotenv
requests
matplotlib
seaborn
numpy
scikit-learn
patsy
jupyter
